requirements.txt
aiofiles==25.1.0
APScheduler==3.11.1
cffi==2.0.0
cryptography==46.0.3
pycparser==2.23
tzlocal==5.3.1
argon2-cffi==25.1.0
bcrypt==5.0.0